# Misc. render optimizations