--- a/core/shaders/point.vs
+++ b/core/shaders/point.vs
@@ -44,7 +44,7 @@ varying vec4 v_color;
 varying vec2 v_texcoords;
 #ifdef TANGRAM_TEXT
 varying float v_sdf_threshold;
-varying float v_sdf_scale;
+varying float v_sdf_pixel;
 #endif
 varying float v_alpha;
 
@@ -71,16 +71,17 @@ void main() {
 #ifdef TANGRAM_TEXT
     vec2 vertex_pos = UNPACK_POSITION(a_position);
     v_texcoords = UNPACK_TEXTURE(a_uv);
-    v_sdf_scale = a_scale / 64.0;
+    float sdf_scale = a_scale / 64.0;
+    v_sdf_pixel = 0.5 / (u_max_stroke_width * sdf_scale);
 
     if (u_pass == 0) {
         v_sdf_threshold = 0.5;
     } else {
         if ((a_stroke.a > 0.0)) {
-            float stroke_width_2;
-            stroke_width_2 = ((a_stroke.a * u_max_stroke_width) * (0.5 / u_max_stroke_width));
-            stroke_width_2 = (stroke_width_2 / v_sdf_scale);
-            v_sdf_threshold = max ((0.5 - stroke_width_2), 0.0);
+            float stroke_width;
+            stroke_width = ((a_stroke.a * u_max_stroke_width) * (0.5 / u_max_stroke_width));
+            stroke_width = (stroke_width / sdf_scale);
+            v_sdf_threshold = max ((0.5 - stroke_width), 0.0);
             v_color.xyz = a_stroke.xyz;
         } else {
             v_alpha = 0.0;