--- a/core/src/gl/vao.cpp
+++ b/core/src/gl/vao.cpp
@@ -7,7 +7,7 @@
 
 namespace Tangram {
 
-void Vao::initialize(RenderState& rs, ShaderProgram& _program, const std::vector<std::pair<uint32_t, uint32_t>>& _vertexOffsets,
+void Vao::initialize(RenderState& rs, ShaderProgram& _program, const VertexOffsets& _vertexOffsets,
                VertexLayout& _layout, GLuint _vertexBuffer, GLuint _indexBuffer) {
 
     m_glVAOs.resize(_vertexOffsets.size());