--- a/core/src/gl/vao.h
+++ b/core/src/gl/vao.h
@@ -10,11 +10,13 @@ class RenderState;
 class ShaderProgram;
 class VertexLayout;
 
+using VertexOffsets = std::vector<std::pair<uint32_t, uint32_t>>;
+
 class Vao {
 
 public:
 
-    void initialize(RenderState& rs, ShaderProgram& _program, const std::vector<std::pair<uint32_t, uint32_t>>& _vertexOffsets,
+    void initialize(RenderState& rs, ShaderProgram& _program, const VertexOffsets& _vertexOffsets,
                     VertexLayout& _layout, GLuint _vertexBuffer, GLuint _indexBuffer);
     bool isInitialized();
     void bind(unsigned int _index);