- Use vaos for first batch of dynamic buffer mesh (looking at the trace, vertex bytes offset is different only for specific use cases; when a mesh is too big, or when a texture is specific to a mesh through a bitmap marker. Since the byte vertex offset cannot be captured by the vertex array object, only use it when this offset is 0, which makes it being used most of the time)
- Optimize costly shaders (Shader program 60 on the screenshot 3.59ms -> 1.55ms): 
   - Offload some of costly shader work from fragment to vertex shader
   - Pass part of it through glsl optimizer and adapt to something still human readable
- Configure `GLKView` to disable stencil and use no msaa on iOS

For the same area on walkabout the frame rate time changed from ~6.3ms to ~3.8ms on iPhone 6S Plus:

<img width="813" alt="screen shot 2017-03-17 at 11 52 20 am" src="https://cloud.githubusercontent.com/assets/7061573/24051380/45c06f20-0b08-11e7-99c2-99db0ccdc710.png">
<img width="817" alt="screen shot 2017-03-17 at 11 52 37 am" src="https://cloud.githubusercontent.com/assets/7061573/24051382/478c7a88-0b08-11e7-8d5b-19f1ba17fe80.png">