@hjanetzek 2017-03-17
@karimnaaji do you have a timing just without msaa compared to master? I'm just curious.

DynamicMesh Vaos  :+1: