@karimnaaji 2017-03-17
@hjanetzek a timing with master with only msaa disabled in same environment:
<img width="811" alt="screen shot 2017-03-17 at 3 15 52 pm" src="https://cloud.githubusercontent.com/assets/7061573/24059232/a89d8dbe-0b24-11e7-9dd1-aab03cae61c9.png">