@karimnaaji 2017-03-17
@hjanetzek here are some more ideas that we can use for later improvements and similar tweaks:
- Allow to increase the size of the glyph texture atlas based on the pixel density, right now they get filled up quite quickly on higher density screen display. The only drawback that I faced when changing that is that we need a way to invalidate all texture atlases that could have been created, since pixel ratio can change at runtime when moving the window between screens, and it messes up all UVs previously created. This would reduces the number of vertex buffer objects since they are tight to their texture atlas. I tested 512x512 which reduces the number of draw calls for icons/text by 2 on those devices.
- Passing more shader within the glsl-optimizer and manually tweak them while keeping them human readable (I have not looked at polygon shader so far).
- Use https://www.shadertoy.com/view/4ldSD2 on sdf instead of smoothstep, I have not yet looked at the visual difference with that so I haven't incorporated it, but the amount of instructions generated is effectively reduced as expected.