@hjanetzek 2017-03-20
512x512 could be a sensible default for glyph textures. One reason to keep them smaller was that an atlas gets only discarded and reused when no more meshes reference them. So when one makes a world tour visiting all map regions with different scripts one may hold on to large atlases with very few referenced glyphs.