@karimnaaji 2017-03-22
For comparison @nvkelso, @sensescape  here are some screenshots for anti-aliasing, we will still leave that as an option on the iOS framework but we might turn it off by default.

enabled:
![img_0007](https://cloud.githubusercontent.com/assets/7061573/24222691/4b650f7e-0f29-11e7-90f2-2fcc08d760bd.jpg)
disabled:
![img_0008](https://cloud.githubusercontent.com/assets/7061573/24222692/4c4d6fe4-0f29-11e7-90e0-fab77b16af49.jpg)