@nvkelso 2017-03-22
Were these generated on an iPhone "+" sized phone?

On Wed, Mar 22, 2017 at 3:04 PM, Karim Naaji <notifications@github.com>
wrote:

> For comparison @nvkelso <https://github.com/nvkelso>, @sensescape
> <https://github.com/sensescape> here are some screenshots for
> anti-aliasing, we will still leave that as an option on the iOS framework
> but we might turn it off by default.
>
> enabled:
> [image: img_0007]
> <https://cloud.githubusercontent.com/assets/7061573/24222691/4b650f7e-0f29-11e7-90f2-2fcc08d760bd.jpg>
> disabled:
> [image: img_0008]
> <https://cloud.githubusercontent.com/assets/7061573/24222692/4c4d6fe4-0f29-11e7-90e0-fab77b16af49.jpg>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/tangrams/tangram-es/pull/1385#issuecomment-288554399>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA0EO0q8HVPA0L4cL0vETeKW1NJFEvx1ks5roZsFgaJpZM4MgxR_>
> .
>