@matteblair 2017-03-28
This probably isn't related to the changes here, but I'm noticing that in the iOS demo app a screenshot is being captured whenever "view complete" occurs, even when the call to ` captureScreenshot` is removed. 

Based on the logic for capturing screenshots in `TGMapViewController` (https://github.com/tangrams/tangram-es/blob/master/platforms/ios/src/TangramMap/TGMapViewController.mm#L871-L881) it seems that if a delegate is present that responds to `didCaptureScreenshot` then a screenshot will be captured every time "view complete" occurs (whether a screenshot was requested or not). That doesn't seem right.