@karimnaaji 2017-03-29
That effectively doesn't seem right. There is a missing `self->captureFrameWaitForViewComplete = false` to reset ths flag on that block.