@matteblair 2017-03-29
Something here is causing consistent, immediate crashes in the android demo on our Galaxy S3. The call stack isn't consistent, sometimes it's in our code, sometimes it's in the GL driver, but it's always a bad deallocation for the address `deadbaad`. Example:
```
F/libc    ( 5878): @@@ ABORTING: INVALID HEAP ADDRESS IN dlfree
F/libc    ( 5878): Fatal signal 11 (SIGSEGV) at 0xdeadbaad (code=1)
I/DEBUG   ( 5063): *** *** *** *** *** *** *** *** *** *** *** *** *** *** *** ***
I/DEBUG   ( 5063): Build fingerprint: 'samsung/d2vl/d2can:4.0.4/IMM76D/I747MVLALE8:user/release-keys'
I/DEBUG   ( 5063): pid: 5878, tid: 5891  >>> com.mapzen.tangram.android <<<
I/DEBUG   ( 5063): signal 11 (SIGSEGV), code 1 (SEGV_MAPERR), fault addr deadbaad
I/DEBUG   ( 5063):  r0 deadbaad  r1 00000001  r2 40000000  r3 00000000
I/DEBUG   ( 5063):  r4 00000000  r5 00000027  r6 400b8526  r7 400c7574
I/DEBUG   ( 5063):  r8 00c90b90  r9 00000009  10 00000008  fp 009d0ca8
I/DEBUG   ( 5063):  ip ffffffff  sp 5dccd4b0  lr 40099525  pc 400958ac  cpsr 60000030
I/DEBUG   ( 5063):  d0  6772612064696c75  d1  766e49203232206d
...
I/DEBUG   ( 5063):          #00  pc 000178ac  /system/lib/libc.so
I/DEBUG   ( 5063):          #01  pc 0001373a  /system/lib/libc.so
I/DEBUG   ( 5063):          #02  pc 00015a74  /system/lib/libc.so (dlfree)
I/DEBUG   ( 5063):          #03  pc 000160ec  /system/lib/libc.so (free)
I/DEBUG   ( 5063):          #04  pc 000477e0  /system/lib/egl/libGLESv2_adreno200.so (rb_vbo_free)
I/DEBUG   ( 5063):          #05  pc 0003e50c  /system/lib/egl/libGLESv2_adreno200.so (rb_delete_detach_vbo_list)
I/DEBUG   ( 5063):          #06  pc 0005c1de  /system/lib/egl/libGLESv2_adreno200.so (leia_perform_resolve)
I/DEBUG   ( 5063):          #07  pc 0003f25a  /system/lib/egl/libGLESv2_adreno200.so (rb_resolve)
I/DEBUG   ( 5063):          #08  pc 00041760  /system/lib/egl/libGLESv2_adreno200.so (rb_surface_swap)
I/DEBUG   ( 5063):          #09  pc 00030882  /system/lib/egl/libGLESv2_adreno200.so (gl2_surface_swap)
I/DEBUG   ( 5063):          #10  pc 00026bb4  /system/lib/egl/libGLESv2_adreno200.so (oglSwapBuffer)
I/DEBUG   ( 5063):          #11  pc 0000da5c  /system/lib/egl/libEGL_adreno200.so (qeglDrvAPI_eglSwapBuffers)
I/DEBUG   ( 5063):          #12  pc 000055f8  /system/lib/egl/libEGL_adreno200.so (eglSwapBuffers)
I/DEBUG   ( 5063):          #13  pc 0000cc80  /system/lib/libEGL.so (eglSwapBuffers)
I/DEBUG   ( 5063):          #14  pc 0004c2aa  /system/lib/libandroid_runtime.so
I/DEBUG   ( 5063):          #15  pc 0001fbb0  /system/lib/libdvm.so (dvmPlatformInvoke)
I/DEBUG   ( 5063):          #16  pc 0005a734  /system/lib/libdvm.so (_Z16dvmCallJNIMethodPKjP6JValuePK6MethodP6Thread)
```
On the master branch, the android demo is slow, but stable on the S3.