@bcamper 2017-03-29
> a bad deallocation for the address `deadbaad`

:metal:💀 ⛔️