@karimnaaji 2017-03-29
Ah no... think we've come up with that in the past which is why we haven't used them to render text, I wonder if there is not a bug in the driver whenever a shared index buffer is used between different VAOs. That would be the only difference with usage of GL resources from master that I would see in this branch. Nothing comes to mind when I try to think what we could be going wrong.