@matteblair 2017-03-29
More complete log from another occurrence of the crash: https://pastebin.com/0VL6Se5s