@matteblair 2017-03-29
The logs indicate that tangram thinks VAOs are supported here. I wouldn't be shocked if we're hitting a bug in the driver, but we can at least narrow down the cause a bit first. 

I'll try this on other devices and see how isolated the crash is.